Modularizando e tornando mais legivel
